CLEANUP: Lots of Win64 types

+++ src/ker/assoc.c
@@ -30,7 +30,7 @@
 static int	host_handles;		/* # handles */
 static long	itf_symbols;		/* # symbols */
 
-#define SizeOfSymbol ((int)(long)(&(((PceITFSymbol)NULL)->handle[host_handles])))
+#define SizeOfSymbol ((size_t)(intptr_t)(&(((PceITFSymbol)NULL)->handle[host_handles])))
 
 		/********************************
 		*            SYMBOLS		*
